+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - updated energy storage power station reportings
 - changed meter search box size to small
+- updated `myems_system_db`.`tbl_energy_storage_containers_power_conversion_systems` in database
 ### Fixed
 - fixed typo in myems-web routes
 ### Removed